backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.12.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
supabase==2.0.3
python-multipart==0.0.6
websockets==12.0
google-generative-ai==0.2.0 